+++ src/main/java/galaxy/generator/Weighted.java
@@ -2,6 +2,6 @@ package galaxy.generator;
 
 public interface Weighted {
 
-	float weight();
+	double weight();
 
 }
